refactor(#444): split innovative.ts (913 lines) into innovativeData/ modules
